requirements.txt
openai==1.3.5
python-dotenv==1.0.0
langchain==0.0.340
streamlit==1.28.2
wikipedia==1.4.0
numexpr==2.8.7